@@ -3,83 +3,50 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
+    "dev": "vite",
+    "serve": "vite",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview",
+    "lint": "eslint --ext ts,vue --ignore-path .gitignore ./src/",
+    "prettier": "prettier --check ./src/",
+    "fix": "run-s fix:lint fix:prettier",
+    "fix:lint": "eslint --fix --ext ts,vue --ignore-path .gitignore ./src/",
+    "fix:prettier": "prettier --write ./src/",
     "test": "jest --maxWorkers=2",
     "test-unit": "jest --maxWorkers=2 --watch"
   },
   "dependencies": {
-    "konva": "^6.0.0",
-    "uikit": "^3.9.4",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.3"
+    "@vue/eslint-config-typescript": "^10.0.0",
+    "eslint": "^8.14.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-vue": "^8.7.1",
+    "konva": "^8.3.6",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.6.2",
+    "tslib": "^2.4.0",
+    "uikit": "^3.14.0",
+    "vue": "^3.2.33",
+    "vue-router": "^4.0.14"
   },
   "devDependencies": {
-    "@types/jest": "^25.2.3",
-    "@types/lodash": "^4.14.178",
+    "@babel/plugin-transform-modules-commonjs": "^7.17.9",
+    "@babel/preset-env": "^7.17.10",
+    "@types/jest": "^27.5.0",
+    "@types/lodash": "^4.14.182",
+    "@types/path-browserify": "^1.0.0",
     "@types/source-map-support": "^0.5.4",
     "@types/uikit": "^3.3.2",
-    "@typescript-eslint/eslint-plugin": "^2.34.0",
-    "@typescript-eslint/parser": "^2.34.0",
-    "@vue/cli-plugin-eslint": "^4.3.0",
-    "@vue/cli-plugin-typescript": "^4.3.0",
-    "@vue/cli-service": "^4.3.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/eslint-config-typescript": "^5.1.0",
-    "eslint": "^6.0.0",
-    "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-vue": "^6.2.2",
-    "jest": "^26.6.3",
+    "@vitejs/plugin-vue": "^2.3.1",
+    "babel-jest": "^27.5.1",
+    "canvas": "^2.9.1",
+    "jest": "^27.5.1",
+    "jest-localstorage-mock": "^2.4.21",
     "lodash": "^4.17.21",
-    "prettier": "^1.19.1",
+    "path-browserify": "^1.0.1",
     "source-map-support": "^0.5.21",
-    "ts-jest": "^26.5.6",
-    "typescript": "~3.8.3",
-    "vue-template-compiler": "^2.6.14"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended",
-      "@vue/typescript/recommended",
-      "@vue/prettier",
-      "@vue/prettier/@typescript-eslint"
-    ],
-    "parserOptions": {
-      "ecmaVersion": 2020
-    },
-    "rules": {}
-  },
-  "jest": {
-    "verbose": true,
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js"
-    ],
-    "moduleNameMapper": {
-      "^@(/.*)$": "<rootDir>/src/$1"
-    },
-    "transform": {
-      "^.+\\.(ts|tsx)$": "ts-jest"
-    },
-    "globals": {
-      "ts-jest": {
-        "tsconfig": "tsconfig.json"
-      }
-    },
-    "testMatch": [
-      "**/test/**/*Spec.+(ts|tsx|js)"
-    ]
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "ts-jest": "^27.1.4",
+    "typescript": "^4.6.4",
+    "vite": "^2.9.7",
+    "vue-tsc": "^0.34.11"
+  }
 }
