@@ -7,9 +7,9 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent } from "vue";
 
-export default Vue.extend({
-  name: "App"
+export default defineComponent({
+  name: "App",
 });
 </script>
