@@ -16,17 +16,17 @@
     </div>
     <div id="save-key-phrase" class="save-item-container">
       <input
+        v-model.trim="inputKeyPhrase"
         type="text"
         max="50"
         class="uk-input uk-form-small"
-        v-model.trim="inputKeyPhrase"
         list="phrase-history"
       />
       <datalist id="phrase-history">
         <option v-for="p in phraseHistory" :key="p">{{ p }}</option>
       </datalist>
     </div>
-    <div class="save-item-container" style="display: flex;">
+    <div class="save-item-container" style="display: flex">
       <div class="save-output-btn btn-red" @click="submitPhrase">OK</div>
       <div class="save-output-btn btn-gray" @click="resetForm">RESET</div>
     </div>
@@ -34,22 +34,23 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent } from "vue";
 
 type DataType = {
   keyPhrase: string;
   phraseHistory: string[];
 };
 
-export default Vue.extend({
+export default defineComponent({
   name: "EditorOption",
+  emits: ["save"],
   data(): DataType {
     const phraseHistory: string[] = JSON.parse(
       localStorage.getItem("keyPhrases") || "[]"
     );
     return {
       keyPhrase: "",
-      phraseHistory
+      phraseHistory,
     };
   },
   computed: {
@@ -59,8 +60,8 @@ export default Vue.extend({
       },
       set(keyPhrase: string) {
         this.keyPhrase = keyPhrase;
-      }
-    }
+      },
+    },
   },
   methods: {
     submitPhrase() {
@@ -69,7 +70,7 @@ export default Vue.extend({
       const phrasesLengthMax = 10;
       if (this.phraseHistory.includes(keyPhrase)) {
         // keyPhraseを先頭に置く
-        const arr = this.phraseHistory.filter(x => x !== keyPhrase);
+        const arr = this.phraseHistory.filter((x) => x !== keyPhrase);
         arr.unshift(keyPhrase);
         this.phraseHistory = arr;
       } else {
@@ -84,7 +85,7 @@ export default Vue.extend({
 
     resetForm() {
       this.keyPhrase = "";
-    }
-  }
+    },
+  },
 });
 </script>
