@@ -1,42 +1,43 @@
 <template>
   <input
+    v-model="inputValue"
     type="number"
     step="0.01"
     class="editor-speed-piece"
-    v-model="inputValue"
     :style="{ top: top + 'px' }"
   />
 </template>
 
 <script lang="ts">
-import Vue, { PropType } from "vue";
+import { defineComponent, PropType } from "vue";
 import { SpeedType } from "../../model/Speed";
 import toPx from "./helper/toPx";
-export default Vue.extend({
+export default defineComponent({
   name: "SpeedPiece",
   model: {
     prop: "propValue",
-    event: "changeSpeed"
+    event: "changeSpeed",
   },
   props: {
-    isReverse: Boolean,
-    position: Number,
-    propValue: Number,
-    type: { type: String as PropType<SpeedType> }
+    isReverse: { type: Boolean, required: true },
+    position: { type: Number, required: true },
+    propValue: { type: Number, default: 1 },
+    type: { type: String as PropType<SpeedType>, required: true },
   },
+  emits: ["changeSpeed"],
   computed: {
     top(): number {
-      return toPx(this.position, this.isReverse) + 15;
+      return toPx(this.position || 0, this.isReverse) + 15;
     },
     inputValue: {
       get(): number {
-        return this.propValue;
+        return this.propValue || 1;
       },
       set(inputValue: string) {
         const value = Number(inputValue);
         this.$emit("changeSpeed", value);
-      }
-    }
-  }
+      },
+    },
+  },
 });
 </script>
