@@ -5,13 +5,13 @@
 </template>
 
 <script lang="ts">
-import Vue from "vue";
+import { defineComponent } from "vue";
 import ConfigureMain from "@/components/configure/ConfigureMain.vue";
 
-export default Vue.extend({
+export default defineComponent({
   name: "Configure",
   components: {
-    ConfigureMain
-  }
+    ConfigureMain,
+  },
 });
 </script>
