@@ -15,7 +15,7 @@ export class SpeedPieceService {
   // 速度変化コマの存在判定
   hasSpeedPiece(page: number, position: number): boolean {
     return this.scoreData.scores[page - 1].speeds.some(
-      speed => speed.position === position
+      (speed) => speed.position === position
     );
   }
 
@@ -24,7 +24,7 @@ export class SpeedPieceService {
     this.scoreData.scores[page - 1].speeds.push({
       position,
       value,
-      type
+      type,
     });
   }
 
@@ -32,7 +32,7 @@ export class SpeedPieceService {
   remove(page: number, position: number) {
     this.scoreData.scores[page - 1].speeds = this.scoreData.scores[
       page - 1
-    ].speeds.filter(speed => speed.position !== position);
+    ].speeds.filter((speed) => speed.position !== position);
   }
 
   // 速度変化コマの描画
@@ -49,7 +49,7 @@ export class SpeedPieceService {
       fill: color,
       x: this.editorWidth + radius,
       y: toPx(position, this.isReverse),
-      id: `speed-${position}`
+      id: `speed-${position}`,
     });
 
     notesLayer.add(triangle);
