@@ -1,13 +1,10 @@
-import Vue from "vue";
 import Editor from "./pages/Editor.vue";
 import Start from "./pages/Start.vue";
 import Configure from "./pages/Configure.vue";
-import Router from "vue-router";
+import { createRouter, createWebHashHistory } from "vue-router";
 import { CustomKeyKind } from "./model/KeyKind";
 
-Vue.config.productionTip = false;
-
-Vue.use(Router);
+//Vue.config.productionTip = false;
 
 type PropsType = {
   loadScoreDataStr: string;
@@ -15,12 +12,13 @@ type PropsType = {
   selectedKey: CustomKeyKind;
 };
 
-export default new Router({
+export default createRouter({
+  history: createWebHashHistory(),
   routes: [
     {
       path: "/",
       name: "start",
-      component: Start
+      component: Start,
     },
     {
       path: "/editor",
@@ -28,16 +26,16 @@ export default new Router({
       component: Editor,
       props(routes): PropsType {
         return {
-          loadScoreDataStr: routes.params.scoreData,
-          loadMusicUrl: routes.params.musicUrl,
-          selectedKey: (routes.params.key || "7") as CustomKeyKind
+          loadScoreDataStr: routes.params.scoreData as string,
+          loadMusicUrl: routes.params.musicUrl as string,
+          selectedKey: (routes.params.key || "7") as CustomKeyKind,
         };
-      }
+      },
     },
     {
       path: "/configure",
       name: "configure",
-      component: Configure
-    }
-  ]
+      component: Configure,
+    },
+  ],
 });
